requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-decouple==2.1
django-haystack==3.2.1
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pypdf2==3.0.1
pysolr==3.9.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
